5.RAG/requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0